poc: Fix bug in FlpGeneric triggered by multiple gadgets
Previous to this commit, `gadget_poly_len` was the same as the total gadget chunk length (proof length for a one-gadget circuit), which was off by the gadget wire seed length (gadget arity). The tests passed previous to c1548a7 because all of the tested circuits had only a single gadget. As the wire seed is before the polynomial, expecting too many field elements of the polynomial coefficient only attempted to index off the end of the proof, which python truncated to the correct length. As shown by the test in c1548a7, when there was more than one gadget, the code prior to this commit would read part of the second gadget's wire seed into the polynomial of the first. This commit corrects the number of field elements to read for the gadget polynomial.