poc: Remove public param and simplify verify param
There are currently no VDAFs that use the public parameter. As noted
noted in #18 and #19, this creates friction for applications like DAP
(formerly known as PPM) that require frequent rotation of the
verification parameter shared by the Aggregators. This change addresses
this issue by dropping the public parameter from the syntax.

For our current set of VDAFs (Prio3 and Poplar1), the verification
parameter held by an Aggregator is basically a shared secret and some
sort of identifier of its role (aggregator 0, aggregator 1, ...). This
change codifies these two values in the syntax by dropping the
verification parameter and replacing it with a shared "verification key"
and an "aggregator ID". This allows us to replace the `setup()`
procedure with simply generating a shared secret, paving the way for a
solution for #18 that doesn't require breaking the abstraction boundary.