define "domain separation tag"
@jbr pointed out that the `dst` variables that occur in a few VDAF
instantiations are a bit inscrutable. This commit adds `DST` associated
constants to the DAF and VDAF definitions which are then used in PRGs.
The intent is to (1) provide a clear definition of a DST and (2) reduce
the number of places where the string `vdaf-00 prio3` has to be updated
to `vdaf-01 prio3`.