Specify encoding of aggregate shares
Removes the associated type `AggShare` from `Vdaf` and `Daf` and
modifies the syntax to treat the aggregate share as an opauqe byte
string. This is consistent with input shares, which, like the aggregate
share, are ncessarily transmitted over the network.

This partially addresses #58. In the future we may want to define a type
for every quantity and have that type implement a type whereever
applicable. Or we might decide not to be prescriptive about encoding at
all.

Co-authored-by: Phillipp Schoppmann <schoppmp@users.noreply.github.com>