poc: Define `FftField` for FFT-friendly fields
Allow fields to be defined that aren't suitable for FFT. Not all of our
applications will need this.

While at it, define two fields that will be used for the Poplar IDPF
spec. One of these fields, GF(2^255 - 19), is too large for the current
implementation of `next_power_of_2()`. Adjust this accordingly.