Update spec spec to match reference code
All planned changes to Prio3 have been implemented in the reference
code; propagate these changes to the document itself. There are also a
few changes for Poplar1 to pick up as well.

Changes to the reference code were required in order to make it render
nicely in the dcocument. However the algorithms themselves have not
changed.

This change also breaks up the auxiliary functions sections into
subsections: One for joint randomness computation, another for
serialization, and a new subsection for enumerating the constants.