Use SHA-3 to instantiate the PRG and improve domain separation
Replace `PrgAes128` with a new scheme, `PrgSha3`, that uses SHA-3 in the
cSHAKE128 mode of operation. Existing analysis models this object as a
random oracle, and SHA-3 is a safer way to instantiate it.

Whenever the PRG is used to map a seed and some context to a finite
domain, e.g., derive a fresh seed or a vector of field elements, the
analysis models this function as an independent random oracle. One
consequence of this is that we need to take care to ensure domain
separation for each distinct usage of PRG. In particular, we need to
format the inputs to the random oracle so that they are decodable.

To make this easier, we split the current "info string" into two
components: The "context string" and the "binder string". The context
string encodes the document version, algorithm, and some usage
information; all of which was previously packed into the prefix of the
info string. The "binder string" is an optional ephemeral artifact that
we want to bind ot the output, e.g., the measurement shares in Prio3.

We set cSHAKE128's "customization string" to the context string. The
seed is set to the start of the "main input" of cSHAKE128; and the
remainder is the binder string. This is useful because the binder string
may have arbitrary length; because the seed is always fixed length, it
can always be decoded from the main input.