Poplar1: Split PRG for leaf/inner sketch offsets
Use a different `Prg` instance for the inner and leaf nodes when
generating shares of the offsets for the sketch computation (i.e., the
`(a, b, c)` triples). This change is meant to reduce implementation
complexity. In particular, for performance reasons it is desirable to
buffer the seed stream output by the `Prg` prior to doing rejection
sampling. The buffer size is picked based the size of the field
elements; changing the field size is fiddly, as this requires us to copy
the unused part of the buffer to the beginning when we wrap around the
end. For details, see:
https://github.com/divviup/libprio-rs/pull/429