Have each `Vdaf` specify its nonce size
Add associated parameter `NONCE_SIZE` to the `Vdaf` API that determines
the nonce size for that VDAF. For `Prio3` and `Poplar1`, pick a nonce
length of 16 bytes.

To make the API self-documenting, we use Python3's type hints to
indicate that the `Vdaf` member functions
(`measurement_to_input_shares()` and `prep_init()` in particular) expect
a nonce of length `NONCE_SIZE`. The intent is to indicate that only
nonces of that length are allowed. This change also adjusted the type of
hint for `verify_key` accordingly.

Co-authored-by: David Cook <dcook@divviup.org>