Revise security considerations for verification key
Previously we had RECOMMENDED that the verification key be chosen only
after the reports are generated. This guidance was meant to provide
"defense-in-depth" for robustness, as it would make it impossible for
a malicious Aggregator to reveal the key to Clients prior to submission.

Based on the analysis of Prio3 in {{DPRS23}}, we cannot rule out attacks
on privacy if the verification key can be picked by the attacker *and8
the verification key can depend on the random coins of the experiment.
If we wish to allow the attacker to control the verification key, then
we must force the attacker to "commit" to the key prior to generating
reports. (We also require the nonce to be picked at random.)

Remove this RECOMMENDATION and add that the Aggregators MUST commit to
the key before processing reports. Add a suggestion about how task
configuration in DAP-03 might be implemented to support this.