Use littlen-endian byte-order for fields
Add methods to_le_bytes/from_le_bytes for encoding/decoding unsigned
integers as little-endian bytes and use these for field element
serialization. This change is meant to improve alignment of the spec
with implementations, which largely prefer little endian to big endian
for representing field elements.

For consistency, rename I2OSP/OS2IP to to_be_bytes/from_be_bytes.

While at it, check in David's script for generating the PRG rejection
test case. This is not strictly needed, since we could just reverse the
order of the bytes in the test seed, but it will be useful later on.