poc: Plumb coins to `measurement_to_input_shares()`
To improve test vector quality, we will need to be able to specify a
particular sequence of random coins used by the Client to split its
measurement.

Add a paramter to `Daf.measurmeent_to_input_shares()` and
`Vdaf.measurmeent_to_input_shares()` for the random coins. In addition,
have the object specify the size of the byte string it expects.
Similarly for `Idpf`.

Accordingly, remove all calls to `gen_rand()` from `Prio3`, `Poplar1`,
and `IdpfPoplar` and instead use the coins passed to the API.