poc: Don't use class name for test vector name
We currently use the class name to name the file containing teste
vectors for it. We create subclasses to specialize a `Vdaf` (or `Prg`,
`Idpf`, etc.) with a given parameter. It is helpful to have that
subclass have a distinguished name, e.g., `Prio3SumWithShares`. However,
this name is not particularly helpful as a name for the file.

To fix this, add a `test_vec_name` class parameter to each class for
which we generate test vectors and use that as the filename instead of
`__class__`.