Require type definitions for encoded values
The current convention for the `Daf` and `Vdaf` interfaces is to treat
any value that must be written to the network as an opaque byte string
(type `bytes`). This includes the public share, input share, prep share,
prep message, and aggregate share. These values implicitly have type
`bytes` and thus their types don't appear in the interface.

This mixture of explicit and implicit types makes the interface
confusing. At the same time, the current adhoc approach to specifying
encoding of types is hard to understand.

This commit requires each (V)DAF to specify explicit types for all
previously implicitly-`bytes` types. Furthermore, it RECOMMENDS that
concrete (V)DAFs specify encodings for these. Finally, this change:

* Replaces the corresponding types for Prio3 and Poplar1 with more
  explicit types.
* Moves the previous encodings to their own sections and replaces the
  adhoc methods with TLS syntax.
* Replaces some uses of trivial type aliases with the python equivalent.

Co-authored-by: Tim Geoghegan <timg@letsencrypt.org>
Co-authored-by: Brandon Pitman <brandon.pitman@gmail.com>
Co-authored-by: Junye Chen <junyec@apple.com>
Co-authored-by: David Cook <dcook@divviup.org>