Rename PRG to XOF
"Pseudorandom generator" is a misnomer because we can't always treat the
output as pseudorandom (i.e., computationally indistinguishable from
random). We typically require this object to be indifferentiable from a
random oracle.

Clarify this in security considerations and rename PRG to XOF
("eXtendaable Output Function"). This is a term commonly used to
aborb-then-squeeze constructions like SHAKE128.

Co-authored-by: David Cook <dcook@divviup.org>