Prio3: Improve domain separation for multi-proof mode
The query randomness is correlated for different values of
`Prio3.PROOFS`. To provide defense in depth in case the same
verification key is used for different versions, prepend the binder with
`bytes(Prio3.PROOFS)`.

For consistency, do the same for the prove randomness and each Helper's
share of the proofs. Also, do the same for joint randomness in case
`Prio3.PROOFS == 1`.