Remove type aliases, simplify hints, and express bounds in prose
We use Python's type hints to make APIs more self-documenting. However
we abuse hints in a couple of ways:

1. We give aliases to primitive types, e.g., `Vec` for `list`.
2. We try to express fixed-length arrays with hints, e.g.,
   `bytes[Vdaf.NONCE_SIZE]`, but don't (and probably can't) enforce it.
3. We use `Unsigned` for non-negative integers, but don't enforce it.

These abuses have lead to code rot. To address this, we remove the type
aliases altogether, simplify the type hints, and express bounds on
types in prose. In some cases, the bounds are stricter than they were
before.

A few minor changes picked up along the way:

- idpf_poplar.py: `packed_bits` -> `packed` for consistency with
  `encoded` elsewhere.

- vdaf_prio3.py: Remove `PRio3SumVecWithMultiproof.is_recommended()` as
  it is not really used.

- Declare `gen_rand()` to be a CSPRNG.