Enforce type hints for flp.py, flp_generic.py, and tests
The "associated type" pattern we're trying to follow is not enforceable.
For example, we'll do something like this:

```
class Foo:
    Measurement = None
    def bar(self, m: Measurement): pass
```

`mypy` complains because `None` is not a valid type. We instead could do
the following:

```
class Foo:
    Measurement: Any = None
    def bar(self, m: Measurement): pass
```

A concrete sub-class would override this in the natural way:

```
class FooForInt(Foo):
    Measurement = int
    def bar(self, m: int): pass

Foo().bar(23)
Foo().bar("23")
FooForInt().bar(23)
FooForInt().bar("23") # fails
```