Enforce type hints in daf.py, tests/test_dap.py
One notable issue: The signature for `run_daf()` was previously

```
def run_daf(Daf,
            agg_param: Daf.AggParam,
            measurements: list[Daf.Measurement],
            nonces: list[bytes]):
```

We want to treat `Daf` as a generic parameter, but that's not how it's
interpreted by mypy. The first argument, `Daf`, shadows `class Daf`
above, but the type hints here actually refer to the shadowed variable.
To resolve this, simply remove the type hints.