Rename FlpGeneric to FlpBBCGGI19 and IdpfPoplar to IdpfBBCGGI21
The rationale has two parts.

First, the names follow different conventions: "Generic" in the sense
that it applies to any language that can be expressed as an arithmetic
circuit, and "Poplar" simply because it was the IDPF from the Poplar
paper. This is a bit awkward.

Second, neither scheme is named, so the names that apply best are the
papers in which the schemes were described (BBCGGI19 and BBCGGI21
respectively).

The new names only appear in the reference code. IN the document we just
refer to them as the "FLP of {{BBCGGI19}}" and IDPF of {{BBCGGI21}}
respectively.