Prio3: Move field spec from FLP to VDAF
If one wants to use a different field for a circuit other than the
default, they must define a subclass with the `field` parameter
overwritten. This is a bit clunky, and the need to configure the field
is common enough that it's worth reducing the friction.

This commit moves the specification of the field from the validity
circuit up to the VDAF. Accordingly, add a table to each subsection of
{{prio3-instantiations}} that defines the circuit, field, number of
proofs, and XOF.

Other changes:

- Align the FLP syntax with the poc: use `F` as the generic field
  parameter rather than `Field`.

- In the specification of `FlpBBCGGI19`, use `field` instead of `Field`
  as a reference to the field object.

- Remove some outdated references to the poc.