Make the reference code installable via pip
Projects can now install the reference implementation via

```
sage -pip install git+https://github.com/cfrg/draft-irtf-cfrg-vdaf@draft-irtf-cfrg-vdaf-11#subdirectory=poc
```

This avoids the need to hack a submodule into the project's import path.

This change also moves `test_vec` to the root project directory so that
`pip install` can disambiguate this folder from the actual package,
which is called `vdaf_poc` and exists in the same directory. We could
alternatively put the package in a distinguished `src` directory, but
this would require the user to install the package in order to run
tests.