Replace pseudocode for ping-pong API with Python
Add an implementation of the ping-pong API to the reference code. The
pseudocode currently in the draft doesn't compile, so replace it with
the tested, type-checked, and linted Python.

Also, in the draft, we don't need to list the encoding methods used by
the ping-pong API; these should be clear in context.