Resolve open issues for #255
Two questions remain.

First, for consistency with other sections, allow the IDPF to define its
own type for the public share and express the encoding correction words
for {{BBCGG21}} in TLS-syntax.

Second, for consistency with other sections, we express the encoding of
`Poplar1`'s aggregation parameter in TLS-syntax. This is slightly hairy
because of the prefix packing procedure, but this can be factored out
pretty nicely.