Refer to an instance, not the class in msssage encoding sections
To define message encoding, we need to be able to refer to attributes of
the VDAF, e.g., the value of `SHARES`. Previously we referred to the
class in this section, e.g., `Prio3`, but since getting rid of the
class-factor pattern, it is now necessary to refer to an instance of the
class.

To make this clear, replace upper case `Prio3` with lowercase `prio3`,
to imply that we an instance of a sub-class of `Vdaf` rather than the
sub-class itself.

Likewise for `Poplar1` some places.

While at it, fix `Fi` and `Fl` by adding the encoded size.