poc: Add application context string (*)
WIP: We still need to propagate the changes through the draft and update
security considerations, in particular the line about weak parameters
being exploitable across tasks.

Add an application context string, denoted `ctx`, to sharding and
preparation APIs for DAFs and VDAFs. The intent is to use this for
defense-in-depth: by binding execution of a VDAF to its application
(concretely, in DAP we would set `ctx` to the task ID), we hope that an
attack on that application does not translate directly to an attack on
another application using the same parameters.

This change is motivated by the possibility of an offline attack against
the robustness of some Prio3 variants. The attack is as follows: fix an
invalid measurement, then search for measurement shares that derive
joint randomness that causes the circuit to accept. Once these shares
are found, they can be used to corrupt any batch that uses the same
Prio3 parameters.

To reduce the blast radius of this attack, we can bind `ctx` to the
joint randomness derivation, thereby ensuring the offline work done by
the attacker is only useful to break a single application.

This change is slightly more general: wherever we compute a domain
separation tag in either Prio3 or Poplar1, we append `ctx`. This is a
more invasive than necessary, but it's also more conservative.

Finally, we modify XofTurboShake128 and XofFixedKeyAes128 by changing
the length prefix for the domain separation tag from a single byte to
two bytes, thereby accommodating larger application context strings.