Point to test vectors rather than reproduce them
The current plan is to reproduce the test vectors in the draft once
we're ready for RFC. However the value of these test vectors for
implementers is not clear. It would be better to publish
machine-readable test vectors somewhere, point to it from the
draft, and publish a hash of the test vectors in the draft so that
implementers can use to verify their veracity.

This commit implements the following idea: point to a commit hash for
the draft and describe how to consume the test vectors.

This commit also moves the VDAF test vectors to a directory that is easy
to reference. Namely, we remove the sub-directory based on the version
and add a `vdaf` sub-directory.