Add application context string
Propagate changes to the reference code from a recent change that add an
application context string to sharding and preparation. Explain the
intended role of context: disagreement should imply failure to aggregate
a report. However, the affect is different for DAFs and VDAFs, since for
DAFs there is no explicit signal of rejection; disagreement would
manifest as a garbled aggregate result. Point this out.

Also, note in security considerations that this provides a cushion for
weak parameters.

Finally, update the order of the context string in the `Daf` input
parameters to align with the `Vdaf`. The context string comes ahead of
every input except for the VDAF verification key.

Co-authored-by: David Cook <dcook@divviup.org>