Prio3: Improve soundness of reductions (*)
In a validity circuit we may have several values we expect to be 0. To
check that they're all zero, we interpret them as coefficients of a
polynomial and evaluate the polynomial at random point. We do this in
several places:

* Whenever `EVAL_OUTPUT_LEN > 0`
* In the `SumVec` circuit
* In the `MultihotCountVec` circuit
* In the `Histogram` circuit

This induces a soundness error of `n / |F|`, where `n` is the number of
values and $F$ is the finite field. This is significant especially for
larger inputs.

We can reduce this soundness error by reducing the degree of the
polynomial. By the Schwartz-Zippel lemma, when we evaluate a non-zero,
multivariate polynomial with degree `d` at a random point (each variable
gets a random value), the probability that the output is zero is at most
`d / |F|`. For instance, if each value is a different variable, then the
maximum degree is `1`, resulting in a `1 / |F|` soundness error.

However, the more variables the polynomial, the more randomness we need
to produce. This can become a CPU bottleneck for longer inputs.

This commit makes the following changes to Prio3 and its variants:

1. When `EVAL_OUTPUT_LEN > 0`, we include in the query randomness
   `EVAL_OUTPUT_LEN` field elements and compute the reduced output as
   the dot product of this random vector and the outputs. This improves
   the soundness error from `EVAL_OUTPUT_LEN / |F|` to `1 / |F|`.

2. We modify the `SumVec`, `MultihotCountVec` and `Histogram` circuits
   by including in the joint randomness a vector of length
   `GADGET_CALLS[0]`. For the `i`-th gadget call, we generate powers of
   `joint_rand[i]`. This results in a soundness error of `sqrt(length) /
   |F|` when the chunk length is chosen optimally.