Prio3, Poplar1: Set `xof` to `XofTurboShake128`
Previously we imagined making this generic in case we wanted to define
variants with different XOFs. We don't have nay in this document, so fix
XofTurboShake128 as the XOF as early as possible.

While at it, don't overload `Xof` in parameter tables, as this is now
just an abstract base class. Likewise for `Idpf` and `Flp`.