poc: Add an example showcasing FLP with Shamir's
A demonstration of the composition of FLP with Shamir's secret sharing.
This could be the basis of a Prio3 variant that tolerates drop out of an
Aggregator.

Imagine a DAP setup [1] where a Leader wants to be able to run the
protocol with one of two Helpers so that if the first Helper goes
offline, it can continue the computation with the other.

For this to work, the Client still needs to know the identities of the
two Helpers prior to generating its report.

[1] https://github.com/ietf-wg-ppm/draft-ietf-ppm-dap/issues/22