Clarify what we mean by "secure" VDAF execution
"Simulate [some procedure] over an insecure network" is a confusing
concept that we don't unpack very well. All we mean is that `run_vdaf()`
shows ideally how the computation is carried out, but in reality
messages are exchanged over a network the attacker has control of. It's
therefore up to the application to do the stuff described in security
considerations.

While at it, make `run_vdaf()` more readable.