Overview: Disambiguate out shares from agg shares
Comment from Crypto Panel:

> Section 3: Figure 1 shows only "aggregate shares" while the text also
> talks about "output shares". I find the terminology confusing because
> the "output shares" seem to be an intermediate computation step and are
> not an output of the aggregator. Also, Figure 1 could be augmented to
> show all the objects mentioned in the text, such as output shares and
> public shares.

Figure 1 doesn't mention output shares on purpose. There we just want to
show the high level flow of data that is actually transmitted between
parties. Output share aren't transmitted and so aren't included.

I think the confusion is that we don't clearly define "output share",
and it's easy to conflate this with "aggregate share". ("Output" sounds
like the output of the aggregation function.) To resolve this, I added a
paragraph below the input shares -> aggregate shares -> aggregate result
flow that is about preparation.

I also that we have some awkward text that tries to explain the
relationship between input, output, and aggregate shares. Replace this
with a figure, which I think ought to directly address the comment.