prio: Perform polynomial arithmetic in the Lagrange basis. (#585)
* prio: Perform polynomial arithmetic in the Lagrange basis.

The main change is in the FLP, where the (input/output) wire
polynomials must remain in the Lagrange basis. This eliminates
conversions to the monomial basis that are performed by
NTT invocations.

The Prover must send the gadget_poly in the Lagrange basis,
so the Verifier can evaluate it directly in this basis.
Algorithms are described in [Faz25](https://ia.cr/2025/1727).

Updates test vectors.

* Addressing first round of review comments.

* Addressing second round of review comments.

* Addressing commments of review and adding more tests.

* Copying argument locally.